fix: updated patch endpoint so that it takes id as parameter
We don't want the request body to accidentally update the id. This change makes it more clear what doc we are updating. We still need to implement logic in the API to make sure the id is never changed or a new field called "id" is added to the document.